Removed the need for ID field in ByteSpan
The field was used to identify the location of the span in a buffer pool
This change removes the field and now the location is computed using memory adresses.
