+++ src/System.Buffers/src/System/Span.cs
@@ -193,7 +193,7 @@ namespace System
         {
             handle = GCHandle.Alloc(source._array, GCHandleType.Pinned);
             var pinned = handle.AddrOfPinnedObject() + source._index;
-            var byteSpan = new ByteSpan(((byte*)pinned.ToPointer()), source._length, -1);
+            var byteSpan = new ByteSpan(((byte*)pinned.ToPointer()), source._length);
             return byteSpan;
         }
     }
